+++ src/main/java/com/study/simpleboard/service/CommentService.java
@@ -3,8 +3,8 @@ package com.study.simpleboard.service;
 import com.study.simpleboard.common.exception.CustomException;
 import com.study.simpleboard.common.exception.ErrorCode;
 import com.study.simpleboard.dto.CommentCreateDTO;
-import com.study.simpleboard.dto.CommentRequestDTO;
-import com.study.simpleboard.dto.CommentResponseDTO;
+import com.study.simpleboard.dto.request.CommentRequestDTO;
+import com.study.simpleboard.dto.response.CommentResponseDTO;
 import com.study.simpleboard.mapper.CommentMapper;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
